+++ javascript-react-eslint.js
@@ -40,6 +40,7 @@ module.exports = {
         ]
       }
     ],
+    'import/no-default-export': 'error',
     'react/jsx-no-undef': 2,
     'react/jsx-no-duplicate-props': [
       1,
@@ -54,6 +55,7 @@ module.exports = {
     'react/no-unused-prop-types': 2,
     'react/prop-types': 2,
     'react/sort-prop-types': 2,
+    'react/jsx-fragments': ['error', 'element'],
     'react/sort-comp': [
       2,
       {